chore: Update Kubernetes deployment and main.go to include Kafka host
This commit updates the Kubernetes deployment configuration in `deployment.yaml` to include a new environment variable `KAFKA_HOST` with the value "kafka-controller-headless.payetonkawa-prod.svc.cluster.local". Additionally, the `start.go` file is modified to retrieve the `KAFKA_USERNAME` and `KAFKA_PASSWORD` environment variables and log an error if they are not set. This change ensures that the application can connect to Kafka and consume messages in the production environment.