chore: Update Docker build step to use kaniko for building and pushing Docker image
This commit updates the Docker build step in the `.drone.yml` file to use the `gcr.io/kaniko-project/executor:latest` image instead of the `plugins/docker` image. It also sets the `REGISTRY_USERNAME` and `REGISTRY_PASSWORD` environment variables using secrets from the `harbor_username` and `harbor_password` secrets. Additionally, it adds a volume mount for the `kaniko-secret` volume to store the Docker configuration file. The commands are updated to generate the Docker configuration file with the provided credentials and use the `kaniko/executor` to build and push the Docker image to the `registry.germainleignel.com/paye-ton-kawa/auth` repository.