chore: Refactor Kafka consumer to use segmentio/kafka-go library
This commit updates the Kafka consumer in the `kafka/start.go` file to use the `segmentio/kafka-go` library instead of the `confluentinc/confluent-kafka-go/v2` library. It also updates the `ProcessMessage` function signature to match the new library's message type. This change was made to improve the performance and maintainability of the Kafka consumer implementation.