chore: Update Jenkinsfile to use Docker 20.10.7 image and mount docker.sock
This commit updates the Jenkinsfile to use the Docker 20.10.7 image instead of the busybox image for the docker container. It also adds a volume mount for the docker.sock file to enable communication with the host's Docker daemon. These changes were made to improve the Docker-related steps in the pipeline.